requirements.txt
django
django-address
